requirements.txt
requests
beautifulsoup4
anthropic
pydantic>=2.0
python-dotenv
python-telegram-bot>=20.0
apscheduler>=3.10
pytest
